Update macro paths to newest GOV.UK frontend
If you try to run this prototype now it doesn’t work, because the
version of GOV.UK Frontend that you get if running
```shell
npm install govuk-frontend@^3.0.0
```

introduces a breaking change[1] that came with version 3.0.0 of GOV.UK
Frontend.

This repo was using a pre-release version of 3.0.0 in order to get
the new focus styles. However when it was updated to 3.0.0 stable it
wasn’t also updated[2] to respect the folder structure changes that came
with 3.0.0.

This commit updates the paths to the macros to point to the right place.

1. https://github.com/alphagov/govuk-frontend/pull/1458
2. https://github.com/alphagov/focus-state-prototype/commit/22d7f9c6186f11e14b5692b4ba637eff456f0783